Instead of checking for NRF52, check for NRF52_SERIES.
NRF52 apparently (according to line 42 of nrf.h) implies NRF52832_XXAA
which isn't really ideal because this lib (possibly with some tweaks,
which I'm happy to make and send in PRs for) should also work on the
NRF52840.  Setting NRF52 therefore is not what we want (and causes
other build errors in the NRF52840 build).  Instead check for the more
general NRF52_SERIES which is set automatically by the build.